W1: create core/ and move the shared QSO engine into it
First step of the rebuild-plan W1 foundation. Establishes the top-level
core/ directory for framework-agnostic code shared across UIs and (later)
the daemon, and relocates the self-contained QSO engine there.

- git mv ui/qso.ts -> core/qso.ts (pure module, no internal deps); update
  all importers in ui/, ui/web/, and test/.
- Add tsconfig.ui.json + `npm run typecheck` so the core/ + ui/ side has a
  real, repeatable typecheck (now also covering ui/web/**, which the old
  AGENTS.md one-liner did not).

Verified: npm run build, npm run typecheck, and npm test (55) all green.
