+++ core/qso.ts
