+++ ui/web/view-model.ts
@@ -11,7 +11,7 @@ import {
   type DecodeRecord,
   type QsoRecord,
   type TxSlot
-} from "../qso.js";
+} from "../../core/qso.js";
 import type {
   ActiveQsoView,
   CompletedQsoView,
